% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaf_parquet.R
\name{.gaf_parquet_path}
\alias{.gaf_parquet_path}
\title{Path to the parquet file for a given species}
\usage{
.gaf_parquet_path(species)
}
\description{
Path to the parquet file for a given species
}
\keyword{internal}
