% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaf_utils.R
\name{gaf_add_subject_curie}
\alias{gaf_add_subject_curie}
\title{Extract the primary database accession as a CURIE}
\usage{
gaf_add_subject_curie(gaf)
}
\arguments{
\item{gaf}{a tibble as returned by \code{\link{parse_gaf}}.}
}
\value{
\code{gaf} with an additional \code{subject_curie} column prepended.
}
\description{
Combines \code{db} and \code{db_object_id} into a CURIE-style identifier
such as \code{"UniProtKB:P12345"}.  This is the format used in the semsql
\code{gene_go_edge} schema for the \code{subject} column.
}
\examples{
gaf_path <- system.file("extdata", "mini.gaf.gz", package = "GOAnno4Bioc")
gaf <- parse_gaf(gaf_path)
gaf_add_subject_curie(gaf)

}
