% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaf_columns.R
\docType{data}
\name{GAF_COLUMNS}
\alias{GAF_COLUMNS}
\title{GAF 2.2 column names in order}
\format{
An object of class \code{character} of length 17.
}
\usage{
GAF_COLUMNS
}
\description{
A character vector of the 17 column names used in GAF 2.2 files,
in their defined order.  Used internally by \code{\link{parse_gaf}}
to name the columns of the returned tibble.
}
\keyword{datasets}
