% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaf_columns.R
\docType{data}
\name{GAF_EVIDENCE_COMPUTATIONAL}
\alias{GAF_EVIDENCE_COMPUTATIONAL}
\title{Computational evidence codes}
\format{
An object of class \code{character} of length 11.
}
\usage{
GAF_EVIDENCE_COMPUTATIONAL
}
\description{
GAF evidence codes based on computational analysis rather than
direct experiment.
}
\keyword{datasets}
