% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaf_columns.R
\docType{data}
\name{GAF_EVIDENCE_EXPERIMENTAL}
\alias{GAF_EVIDENCE_EXPERIMENTAL}
\title{Experimental evidence codes}
\format{
An object of class \code{character} of length 11.
}
\usage{
GAF_EVIDENCE_EXPERIMENTAL
}
\description{
GAF evidence codes considered to be based on direct experimental
evidence, as opposed to computational or author-inferred codes.
Useful for filtering to high-confidence annotations.
}
\keyword{datasets}
