% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaf_parquet.R
\name{has_gaf_parquet}
\alias{has_gaf_parquet}
\title{Test whether a parquet cache exists for a species}
\usage{
has_gaf_parquet(species = "human")
}
\arguments{
\item{species}{character scalar, e.g. \code{"human"}.}
}
\value{
logical scalar.
}
\description{
Test whether a parquet cache exists for a species
}
\examples{
has_gaf_parquet("human")

}
